feat!: Simplify everything by using interpolations
Dedicate a directory to the macro for `clippy` to like.

Use the `gtk-rs` feature instead of a gtk4-oriented package.

Since the code of the `a_basics` example was reduced,
now it exemplifies builder mode, the rest of
the examples are rotated and all are updated.

The example `z_misc` is renamed to `z_besides` because of its struct.

Give a funny logo to the project.

Add a matrix chat link for questions.