feat!: Replace `'binding` with `@`
Allow using `bindings!()` on arguments or values if `@` is prepended.

An error is thrown when a binding is not consumed and
when trying to consume bindings when there are none.

Now `ref` can reference fields (see the `e_templates` example).

`ref` allowed attributes but is now
exemplified a bit in the `z_besides` example.

Improved memory layout.

Some error messages and example comments were modified.

Basic maintenance is indicated in the `README.md` file.

The code preview in the `README.md` (and thus the `y_readme` example)
is not tab-indented for better viewing in markdown renderers.

General code improvements, especially not using `.into()` in macro code.

Put the logo on the generated documentation.