refactor!: `clone!`, examples and compact strings
Now the `clone!` macro has more features at the
cost of a small breaking change unlikely to suffer.

The code and comments in the examples have been revised and improved.

An example is added to illustrate the possibility
of coercing view closures to function pointers.

Added an alternative example to `components` where no channels are used.

Names are now generated using `format_compact!()` to avoid allocating.