feat!: Integrate with struct fields and set limits
Now builder mode works with structures (see the `z_besides` example).

Due to the above, almost any expression to create an item is
no longer possible, except functions or a single method call.

Now it is possible to edit a struct field using a dot as
`field.method: "args"` and interpolate as `#field.method(#)`.

Due to the above and hygiene, now braces are always required to edit
a struct field like so: `field => { /* edit */ }` (also affects `if`).

It is not possible to bind a field edit (without dot as above) because
it always allows compositing (it was possible due to an oversight).

Now it is true that `'bind:` (with a colon) allows inner conditionals.

Field interpolation does not create extra braces.

Now it is only possible to interpolate in builder mode between
parentheses to avoid a future breaking change on top of this one.

More errors are detected and others are displayed better than before.

However it seems that an error is not displayed correctly when
interpolating with a long path (maybe an update will fix it).

`std::cell::OnceCell` is now used instead of
`once_cell::unsync::OnceCell` in the examples.

Some code and comment fixes in the examples, especially in `z_besides`.