feat: On-demand struct fields and fewer limits
It is possible to declare a struct in the view and generate its fields
according to the items (see examples `f_templates` and `i_no_generics`).

Punctuation is more hygienic as `::default()`, which allows proper
error propagation at the cost of annoying semantic syntax highlighting.

The `conditional` code was moved to `content` for
a simpler and consequently more permissive design.

Due to the above it is possible to bind items or bind in binding or
conditional scopes, which seem to be rare cases (no examples).

The macro does not expand under any error,
except if there are unconsumed bindings.

The assignment or item with a tilde behind it parses attributes.

Consuming bindings with `@var = expr` allows an optional semicolon.

Fixed a bug when using `'back` in builder mode
and not being able to assign when finished.

Some error messages have been changed or
improved and the examples have been updated.