fix!: Naming, hygiene and allow dots in extensions
Name unnamed items according to their `Path` without starting with `_`.

Use the brace span for auto-generated names and `::default()`.

Do not always create a variable when using `'back` in builder mode.

The unconsumed view error is propagated to `view!`.

Empty views are not allowed. They should be allowed, but
there seems to be no way to issue a warning in stable Rust.

Placeholders with content are also not allowed.

Use a tilde in the logo instead of double dots.