feat!: Syntax changes, less code for more features
Composition and extension (v0.5.3 term) do not require parentheses, but
instead extend the property assignment syntax. This allows multiple
composition, even under an extension, as well as `'bind` extensions.
For more details see the `a_basics` example.

Since the new composition syntax uses `@`, the following stop using
`@` to avoid confusion (illustrated in the example `b_reactivity`):
- Arguments that consume bindings do not need to be preceded by `@`.
- For `'bind` to initialize, `#` is appended instead of `@`.
- To consume bindings on a variable, use `'consume` instead of `@`.

The `builder-mode` feature has been removed and the `builder_mode!`
macro has been replaced by `construct!`. And now to change the way in
which the initial content of an item is expanded, a sign is not placed
before the braces, but after them, but there are cases in which the
builder pattern is the default way. Details in the `a_basics` example.

To request a field for a template struct, instead of appending `:`
to the item name, prepend `ref` (equivalent to `pub(self)`) or
another visibility modifier to the name or `mut`, and if the field
type needs to be specified, append `as SomeType` to the name.

To dedicate a scope for editing a field or an inner field,
use `ref the_field { }` instead of `the_field => { }` in an item scope.

Bug fixes:
- A binding scope now behaves the same as a conditional scope.
- `'bind` without braces inherit item attributes.
- A function call cannot be specified in a struct literal item.

Besides:
- Some comment and code fixes in the examples.
- Updated `gtk4` dependency to version 0.7.
- New example: `k_many_params`.
- Improved error messages.