feat: full struct generation and better completion
If `#[view]` is used for an `impl`, it is not necessary to declare a
struct in the view to export items as its fields, but the struct will
be generated with the name of the type to be implemented, including
its attributes and generics. It is also possible to:

- Change the visibility of the struct.
  For example, by typing `#[view(pub)]`.

- Add other fields.
  For example, by typing `#[view(integer: i32, pub float: f32)]`.

- Both at the same time would be
  `#[view(pub, integer: i32, pub float: f32)]`.

Finally, a few instances of useless or incomplete code expansion
were allowed to slightly improve integration with code completion.